Move figure to the middle of the discussion and talk around it.
This framing makes it flow better, so we are talking to a figure while
getting into the details, and giving the BLUF: ASPN or no-ASPN are both
fine as far as pntOS is concerned.

At this top-level, the details of how proprietary gets converted into
ASPN are a detail, so we move that discussion into a Note.