Fix pytest w/ ROS issue
For strange reasons related to pytest's import system, when the ROS
environment is being used, the ROS package launch_testing seems to be
registering itself as a pytest hook, so pytest tries to use it. This
package depends on lark, which we don't have, causing pytest to fail.
For now, just make pytest skip launch_testing.

See #208