Docker: switch to ROS Jazzy container
The CI currently uses Ubuntu 24.04. In order to faciliate the upcoming
ROS tests, we need ROS installed. It's easiest to do this by pulling the
upstream ros:ros-jazzy-base image (which inherits from ubuntu:24.04).

Note that the new container's default ros_entrypoint.sh (which activates
the ROS environment) is being overridden, so the CI will not have the
ROS environment present unless the ROS or ASPN-ROS setup files are
sourced during a run. This is good, as ROS is an optional dependency,
and this allows the CI to easily run tests both with and without it.

See #208