Add ASPN-ROS unit test
When ROS is present, this test ensures that ASPN-ROS and the underlying
ROS transport mechanism is working correctly.

ROS is an optional dependency. So as not to fail for users or CI jobs
that don't have ROS installed and sourced, this pytest is skipped when
running in non-ROS environments.

Before running the test, the job must activate the ASPN-ROS environment
by cloning and sourcing /firehose-outputs/ros_devel/jazzy/setup.bash.
Carson and I considered doing the clone during the build of the docker
container, but we opted against it for two reasons:

1. Since Docker layers using (or following those using) the CI token
   argument do not benefit from caching and the building of the Docker
   image happens in a before_script, the clone would actually happen
   separately for every single job instead of only the jobs using ROS.

2. During local Docker builds (the CI uses tokens instead), secure
   clones using ssh authentication require the --ssh default flag and
   BuildKit. This works with docker and podman, but we are unable to get
   it to work using podman-desktop, limiting macOS support.

Instead, we perform the clone in the container while it's running (only
in the ROS-related jobs). To reproduce this locally, note that the
container will have to be run with ssh key access. But this was already
true because of the clones in the uv sync.

Note that the sourcing of the setup file must happen in the same docker
exec as the test itself since environments don't persist between docker
exec calls.

See #208